scholarmind/requirements.txt
flask==3.1.0
flask-cors==5.0.1
endee==1.0.4
sentence-transformers==3.4.1
PyPDF2==3.0.1
python-dotenv==1.1.0
requests==2.32.3
groq==0.25.0
